+++ tests/Fixtures/App/src/Entity/Product4.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Mtarld\SymbokBundle\Tests\Fixtures\files;
+namespace Mtarld\SymbokBundle\Tests\Fixtures\App\src\Entity;
 
 use Mtarld\SymbokBundle\Annotation\Getter;
+use Mtarld\SymbokBundle\Annotation\Setter;
 
 class Product4
 {
@@ -15,4 +16,10 @@ class Product4
      * @Getter()
      */
     private Product1 $related;
+
+    /**
+     * @var string[]
+     * @Setter(remove=false)
+     */
+    private array $names;
 }